Remove dependency on Pimple.
Pimple 2.x has changed to shared-services-by-default, a decision that’s awesome for a DIC, but not as awesome for a general purpose lazy value bucket like we’re using Pimple for. Since this will likely cause problems with apps which use both Ruler and Pimple, I have removed the Pimple dependency and implemented the Pimple 1.x API directly on Context — using Pimple 2.x-flavored code :)

The extend() method has not been ported to Context, as that seems like a side effect of using a DIC as a base class, more than an explicit feature for the Context. But I’m happy to be proven wrong if someone really needs it.